pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sailing-upwind"
version = "0.1.0"
description = "Simple sailing-upwind optimisation model (from high-school coursework)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
sailing-upwind = "sailing_upwind.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sailing_upwind*"]

[tool.pytest.ini_options]
testpaths = ["tests", "sailing_upwind"]
addopts = "--doctest-modules -v --tb=short"

[tool.coverage.run]
source = ["sailing_upwind"]
